Beginning of the major restructuring: parameterization
The next.train.at function is eliminated as no longer useful.
The sample.arrivals function is completely rewritten to avoid using
next.train.at (which is very inefficient anyway), eliminate all
references to global variables, and add two new features for fuzzing
the input data, a multiplicative gaussian for the passenger counts and
an additive offset for the schedules, to account for potential
inaccuracies in the counting procedure and for passengers who desire
to arrive earlier than the their current train will permit.