fix the inside-out loop in the simulator
The simulator was very slow and also wrong, because the arrival
samples were being generated fresh for every (station, train) pair,
leading to both double counting and undercounting.  The correct thing
was to generate arrival samples once for each station, and then step
row-wise down the schedule to calculate the passengers, which is what
we now do and is also much faster.