run.trials: don't lop off the last station any more
With the change to automatically determine which stations have
boardings data, we no longer need to manually remove the terminal
station from the list -- because it's the terminal no boardings will
ever be recorded and so it will automatically be dropped.

For future work: generate random boardings for infill stations based
on some user-supplied distribution.