refactor a bit more; save all the results
Now writes all the computed results (schedule, 90th percentile
loadings, and number of units required) to a file on every run.  The
sigma of the normal distribution used to fuzz passenger counts is
reduced to .05.  Added a function doit() used to encapsulate the usual
parameters to run.trials() and use it to model all of the scenarios
we've considered so far: { 4, 5, 6 tph } x { all local, short-turn }.

This should be sufficient for a blog post update.