try another "complicated" schedule with changing frequencies
The "rush-hour push" schedule does 3 tph early morning and midday, 6
tph for most of rush hour, and 7.5 tph during the peak period from
8:40 to 9:20.  This loses the benefit of consistent clockface headways
but avoids the need for a 3-unit consist at the very peak, even with
Helsinki-type 75-meter FLIRTs.  (This is good because it means all
units can platform at Yawkey, which is a busy station during peak with
transfers for Longwood and Kendall.  This could be avoided with a
uniform schedule and 100-meter trainsets, but that's a lot of excess
capacity for the times when only one set is needed, and the extra
length might cause problems on other lines where there are platforms
less than 200 m (660 ft) long.)