Fix previous simulation
I specified the offset parameter as 12 because in my pen-and-paper
model I was thinking about the time each train passes Framingham.  But
the simulator cares about the time each train arrives at South
Station, and that's only 3 minutes apart for the same scenario.
Regenerate the results with the correct offset parameter.