A more realistic "zone express" service
This keeps the zone express pattern throughout the day, as the
previous commit, but drops frequency to 4 trains per hour except
during the peak period from 7:00 to 9:30.  At this time period you
could easily do an all-local service, but shaving 9 minutes off the
running time of *all* Worcester trains should reduce equipment
requirements.  The downside is that Newton and Wellesley only get
2-tph off-peak service under this scenario.  You could balance that by
still running 4-tph Framingham service instead but there's not enough
data on travel demand after peak (because of the existing large gaps
in the schedule) to even have an idea how much this would make a
difference.  You could also assign a couple of units to a Riverside
shuttle service that would fill in the gaps in Newton (but leave
Wellesley at 2 tph) since there's no freight traffic.