Add another short-turn schedule
There's been talk about operating short-turn trains from Riverside.
The station there is old and decrepit and unlikely to be usable, but
we don't have travel time estimates to get to the current Green Line
station where a high platform could be built and there's already
parking; the tracks have been lifted but the RoW still exists.  What's
worse, it's a flat junction, and there's no crossover.  Still, it may
be interesting to model if there is demand for increased frequency
during the midday hours.  (You could add a crossover and use the
existing RoW to build a non-revenue layover siding; the nearby Charles
River bridge is three tracks wide, currently used for truck access to
the RoW.)