Update spreadsheets for "zone express" scenario, prune deadwood
This adds loadings and schedules for a "zone express" pattern similar
to what runs today, with 8 tph in the AM peak, and a local/short-turn
pattern at other times (8 tph early shoulder, to get the equipment
into service, and 4 tph middays).  Still needs to be redone with 2018
CTPS passenger counts once these are published.