Simulate the Ag Branch
This is a copy of the train simulator code (yes, I know, it's already
modular, I should just move the library parts into a library and share
them between applications) with additional code to simulate operations
on the Agricultural Branch based on the station siting that I proposed
in an October, 2018, blog post.

Obviously, since we have no way to even guess at passenger counts,
there's no way to use this in the passenger-load simulator, at least
not yet.