go back to 8 minute arrival spacing between local and express
The schedule is much nicer with 10 minute arrival spacing, but that
means trains are only 2 minutes apart at Framingham -- great if you
want to transfer from express to local but not so great operationally
because it means a small schedule slip on the express is much more
likely to cascade to the local.  Going back to 8 minute arrival
spacing increases the Framingham gap to 4 minutes, which is much safer
and allows traffic at the grade crossings at least one clear light
cycle.