model diesel expresses mixed with short-turning local EMUs
One of the concerns expressed by the Rail Vision Advisory Committee
and the MBTA control board is how Regional Rail could be deployed
incrementally.  A model that the staff are considering includes diesel
service to the outer ends of the line mixed with local EMU service at
the inner ends, either as an end state (which would be unacceptable)
or as a transitional state to allow some service to begin while
construction of high platforms and electrification was still ongoing
(or was programmed for later years) at the outer end of a line.

I took the schedule of the current early-morning Worcester express
train P504 and added it to the model, and I was surprised to find that
it actually fits perfectly.  (That's not a reasonable expectation for
other lines, like Providence/Stoughton or the Eastern Route.)  Alon's
model for EMU speeds on the line gives us a 33-minute trip on the
Framingham local, and the current diesel express train takes 39
minutes to do the same run, so an arrival time gap of ten minutes
maintains four minutes of track spacing between the express and the
local.  The express gains only two minutes on the local between Natick
and Yawkey, despite having nine fewer stops, and then starts losing
time again.  So everything fits, without having to build passing
tracks or even more crossovers.