A system map in SVG format
This map is intended to be geographically accurate (at least within
the limits of the Mercator projection and at normal display scale) for
the current commuter rail network and proposed or under-construction
extensions and infill stations.  Routes mainly traced from aerial
photos and adjusted for consistency of presentation and color palette.