Scheduling and simulations for the Providence Line
These schedules assume a transitional situation where the Stoughton
Line is not yet electrified or double-tracked (because significant
station construction is required beyond platforms).  Assumes also that
the Franklin Line is through-routed to the Fairmount for all trips,
and that the Needham Line is operated as a shuttle to avoid fouling
the main line.